--- Source/Moq.Analyzers.Test/ConstructorArgumentsShouldMatchAnalyzerTests.ShouldFailIfClassParametersDoNotMatch.verified.txt
@@ -1,56 +1,0 @@
-﻿Diagnostic 1
-	Id: Moq1002
-	Location: SourceFile(Test0.cs[384..393))
-	Highlight: (1, true)
-	Lines: var mock1 = new Mock<Foo>(1, true);
-	Severity: Warning
-	Message: Parameters provided into mock do not match any existing constructors.
-
-Diagnostic 2
-	Id: Moq1002
-	Location: SourceFile(Test0.cs[428..437))
-	Highlight: (2, true)
-	Lines: var mock2 = new Mock<Foo>(2, true);
-	Severity: Warning
-	Message: Parameters provided into mock do not match any existing constructors.
-
-Diagnostic 3
-	Id: Moq1002
-	Location: SourceFile(Test0.cs[472..480))
-	Highlight: ("1", 3)
-	Lines: var mock3 = new Mock<Foo>("1", 3);
-	Severity: Warning
-	Message: Parameters provided into mock do not match any existing constructors.
-
-Diagnostic 4
-	Id: Moq1002
-	Location: SourceFile(Test0.cs[515..538))
-	Highlight: (new int[] { 1, 2, 3 })
-	Lines: var mock4 = new Mock<Foo>(new int[] { 1, 2, 3 });
-	Severity: Warning
-	Message: Parameters provided into mock do not match any existing constructors.
-
-Diagnostic 5
-	Id: Moq1002
-	Location: SourceFile(Test0.cs[629..659))
-	Highlight: (MockBehavior.Strict, 4, true)
-	Lines: var mock1 = new Mock<Foo>(MockBehavior.Strict, 4, true);
-	Severity: Warning
-	Message: Parameters provided into mock do not match any existing constructors.
-
-Diagnostic 6
-	Id: Moq1002
-	Location: SourceFile(Test0.cs[694..723))
-	Highlight: (MockBehavior.Loose, 5, true)
-	Lines: var mock2 = new Mock<Foo>(MockBehavior.Loose, 5, true);
-	Severity: Warning
-	Message: Parameters provided into mock do not match any existing constructors.
-
-Diagnostic 7
-	Id: Moq1002
-	Location: SourceFile(Test0.cs[758..786))
-	Highlight: (MockBehavior.Loose, "2", 6)
-	Lines: var mock3 = new Mock<Foo>(MockBehavior.Loose, "2", 6);
-	Severity: Warning
-	Message: Parameters provided into mock do not match any existing constructors.
-
