+++ Source/Moq.Analyzers.Test/ConstructorArgumentsShouldMatchAnalyzerTests.ShouldFailWhenConstructorArumentsDoNotMatch.verified.txt
@@ -1,0 +1,32 @@
+﻿Diagnostic 1
+	Id: Moq1002
+	Location: SourceFile(Test0.cs[392..401))
+	Highlight: (1, true)
+	Lines: var mock1 = new Mock<Foo>(1, true);
+	Severity: Warning
+	Message: Parameters provided into mock do not match any existing constructors.
+
+Diagnostic 2
+	Id: Moq1002
+	Location: SourceFile(Test0.cs[436..445))
+	Highlight: (2, true)
+	Lines: var mock2 = new Mock<Foo>(2, true);
+	Severity: Warning
+	Message: Parameters provided into mock do not match any existing constructors.
+
+Diagnostic 3
+	Id: Moq1002
+	Location: SourceFile(Test0.cs[480..488))
+	Highlight: ("1", 3)
+	Lines: var mock3 = new Mock<Foo>("1", 3);
+	Severity: Warning
+	Message: Parameters provided into mock do not match any existing constructors.
+
+Diagnostic 4
+	Id: Moq1002
+	Location: SourceFile(Test0.cs[523..546))
+	Highlight: (new int[] { 1, 2, 3 })
+	Lines: var mock4 = new Mock<Foo>(new int[] { 1, 2, 3 });
+	Severity: Warning
+	Message: Parameters provided into mock do not match any existing constructors.
+
