+++ Source/Moq.Analyzers.Test/ConstructorArgumentsShouldMatchAnalyzerTests.ShouldPassWhenConstructorArgumentsMatch.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
