+++ Source/Moq.Analyzers.Test/Data/ConstructorArgumentsShouldMatch.TestGood.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using Moq;
 
-namespace ConstructorArgumentsShouldMatch;
+namespace ConstructorArgumentsShouldMatch.TestGood;
 
 internal class Foo
 {
@@ -17,21 +17,6 @@ internal class Foo
 
 internal class MyUnitTests
 {
-    private void TestBad()
-    {
-        var mock1 = new Mock<Foo>(1, true);
-        var mock2 = new Mock<Foo>(2, true);
-        var mock3 = new Mock<Foo>("1", 3);
-        var mock4 = new Mock<Foo>(new int[] { 1, 2, 3 });
-    }
-
-    private void TestBadWithMockBehavior()
-    {
-        var mock1 = new Mock<Foo>(MockBehavior.Strict, 4, true);
-        var mock2 = new Mock<Foo>(MockBehavior.Loose, 5, true);
-        var mock3 = new Mock<Foo>(MockBehavior.Loose, "2", 6);
-    }
-
     private void TestGood()
     {
         var mock1 = new Mock<Foo>(MockBehavior.Default);
